refactor: move insights and runway contract types into the api layer
