+++ frontend/src/api/insights/types.ts
@@ -1,5 +1,14 @@
 import type { FxStatus } from '@/api/shared/fx';
-import type { SavedInsightsRangeQualifier, SavedInsightsRangeUnit } from '@/pages/insights/types/range';
+
+// Backend query parameter selecting what the comparison deltas in a range response are computed against
+export type InsightsComparisonPeriod = 'same_length' | 'previous_month' | 'previous_year';
+
+// Relative window units a saved range can step back by, mirrored from the backend schema
+export type SavedInsightsRangeUnit = 'day' | 'week' | 'month' | 'quarter' | 'year';
+
+// How a window is anchored, mirrored from the backend: the current period to date (this), the
+// previous complete period(s) (last), or a rolling window of the last N periods ending today (past)
+export type SavedInsightsRangeQualifier = 'this' | 'last' | 'past';
 
 export interface InsightsPeriodGlanceResponse {
   income: number;
