+++ frontend/src/pages/insights/components/FloatingRangeControl.tsx
@@ -2,9 +2,9 @@ import { useCallback, useEffect, useId, useRef, useState } from 'react'
 import { createPortal } from 'react-dom'
 import { motion, useReducedMotion } from 'motion/react'
 import { Calendar, ChevronDown } from 'lucide-react'
-import type { SavedInsightsRange } from '@/api/insights'
+import type { SavedInsightsRange, SavedInsightsRangeQualifier, SavedInsightsRangeUnit } from '@/api/insights'
 import { joinClassNames } from '@/utils/classNames'
-import type { InsightsRangePreset, SavedInsightsRangeQualifier, SavedInsightsRangeUnit } from '../types/range'
+import type { InsightsRangePreset } from '../types/range'
 import { formatResolvedRangeLabel, getRelativeRangeLabel } from '../utils/range'
 import { RelativeRangeBuilder } from './RelativeRangeBuilder'
 import { SavedRanges } from './SavedRanges'
